+++ release-notes.md
@@ -1,4 +1,5 @@
 Changes since v1.3.1 include
+- don't restore the container versions from the backup
 - fix install issue that prevented SDRplay devices from working out of the box
 - prevent users from inadvertantly locking themselves out of a freshly installed system by restoring from a system with secure-image enabled
 - fix issue with recent iob builds potentially not successfully completing that first install due to CONFIG_NTP_MODE being set to 0